+++ CHANGELOG.md
@@ -8,6 +8,7 @@ with respect to its command line interface and HTTP interface.
 
 ## Unreleased
 ### Fixed
+- Singularity now accepts changes to Startup options in manifest.
 - Off-by-one error with Singularity deploy IDs, fixed in 0.5.9, re-introduced in
   0.5.10. Now includes better tests surrounding edge cases.